Colas-Queue(FIFO)-brach 29

+++ containers/StaQue.h
@@ -9,7 +9,7 @@
     struct Alumno{
         int id;
         string nombre;
-        int nota;
+        float nota;
         string estado;
